requirements.txt
py-sc-kpm==0.4.0
py-sc-client==0.4.0
pyTelegramBotAPI==4.26.0
